Renamed PointsCache to ExecutionPointsCache for clarity
