+++ packages/protocol/analysisManager.ts
@@ -15,7 +15,7 @@ import {
 
 import { addEventListener } from "protocol/socket";
 import { ThreadFront } from "protocol/thread";
-import { preCacheExecutionPointForTime } from "replay-next/src/suspense/PointsCache";
+import { preCacheExecutionPointForTime } from "replay-next/src/suspense/ExecutionPointsCache";
 
 import { client } from "./socket";
 import { MAX_POINTS_FOR_FULL_ANALYSIS } from "./thread/analysis";
