+++ packages/replay-next/components/sources/log-point-panel/HitPointTimeline.tsx
@@ -5,7 +5,7 @@ import { createPortal } from "react-dom";
 import Icon from "replay-next/components/Icon";
 import { SessionContext } from "replay-next/src/contexts/SessionContext";
 import { TimelineContext } from "replay-next/src/contexts/TimelineContext";
-import { imperativelyGetClosestPointForTime } from "replay-next/src/suspense/PointsCache";
+import { imperativelyGetClosestPointForTime } from "replay-next/src/suspense/ExecutionPointsCache";
 import {
   isExecutionPointsGreaterThan,
   isExecutionPointsLessThan,
