+++ packages/replay-next/pages/tests/scopes-inspector.tsx
@@ -3,9 +3,9 @@ import { Suspense, useContext } from "react";
 import Inspector from "replay-next/components/inspector";
 import ScopesInspector from "replay-next/components/inspector/ScopesInspector";
 import Loader from "replay-next/components/Loader";
+import { getClosestPointForTimeSuspense } from "replay-next/src/suspense/ExecutionPointsCache";
 import { getObjectWithPreviewSuspense } from "replay-next/src/suspense/ObjectPreviews";
 import { evaluateSuspense, getPauseIdSuspense } from "replay-next/src/suspense/PauseCache";
-import { getClosestPointForTimeSuspense } from "replay-next/src/suspense/PointsCache";
 import { ReplayClientContext } from "shared/client/ReplayClientContext";
 
 import createTest from "./utils/createTest";
